fix: garbage leaked cloudprovider agentpools (#181)
- background:
In the following scenario, leaked agentpool problem will happen:
1. create machine crd resource, cloudprovider started to create
agentpool resource.
2. delete gpu-provisioner pod, and a new gpu-provisioner pod will be
recreated automatically in the cluster.
3. delete machine crd resource.
4. if cloudprovider spent more than 10min to create agentpool, garbage
collection controller will remove machine crd, and lead to agentpool
leaked.

- solution:
A new agentpool garbage collection(reconcileCloudProviderInstances) is
added for removing leaked agentpools that have no assocsiated nodes and
machines.

+++ pkg/cloudprovider/cloudprovider.go
@@ -98,7 +98,10 @@ func (c *CloudProvider) LivenessProbe(req *http.Request) error {
 
 func (c *CloudProvider) Delete(ctx context.Context, machine *v1alpha5.Machine) error {
 	klog.InfoS("Delete", "machine", klog.KObj(machine))
-	return c.instanceProvider.Delete(ctx, machine.Status.ProviderID)
+	if len(machine.Status.ProviderID) != 0 {
+		return c.instanceProvider.Delete(ctx, machine.Status.ProviderID)
+	}
+	return c.instanceProvider.DeleteByName(ctx, machine.Name)
 }
 
 func (c *CloudProvider) IsMachineDrifted(ctx context.Context, machine *v1alpha5.Machine) (bool, error) {
